moved edit and delete buttons to clients/page.tsx
