+++ src/components/Datatable.tsx
@@ -15,9 +15,6 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table"
-import { Button } from "./ui/button";
-import { Pencil, X } from "lucide-react";
-
 
 interface DataTableProps<TData> {
     columns: ColumnDef<TData, any>[]; 
@@ -33,16 +30,6 @@ interface DataTableProps<TData> {
       columns,
       getCoreRowModel: getCoreRowModel(),
     });
-  
-    const handleEdit = (id: string) => {
-      // Implement edit functionality
-      console.log(`Edit client with id: ${id}`);
-    };
-  
-    const handleDelete = (id: string) => {
-      // Implement delete functionality
-      console.log(`Delete client with id: ${id}`);
-    };
   
     return (
         <div className="rounded-md border">
@@ -76,24 +63,6 @@ interface DataTableProps<TData> {
                         {flexRender(cell.column.columnDef.cell, cell.getContext())}
                       </TableCell>
                     ))}
-                    <TableCell>
-                      <div className="flex space-x-2">
-                        <Button
-                          onClick={() => handleEdit(row.original.id)}
-                          variant="outline"
-                          className="w-8 h-8 p-0 flex items-center justify-center"
-                        >
-                          <Pencil className="w-4 h-4" />
-                        </Button>
-                        <Button
-                          onClick={() => handleDelete(row.original.id)}
-                          variant="destructive"
-                          className="w-8 h-8 p-0 flex items-center justify-center"
-                        >
-                          <X className="w-4 h-4" />
-                        </Button>
-                      </div>
-                    </TableCell>
                   </TableRow>
                 ))
               ) : (
